Split off the node part
